fix: 🐛 i think i fixed the dupes
