+++ apps/antalmanac/src/components/Header/Header.tsx
@@ -1,4 +1,6 @@
 import { AppBar, Box, Stack } from '@mui/material';
+import { useEffect } from 'react';
+import { useNavigate, useSearchParams } from 'react-router-dom';
 
 import Import from './Import';
 import LoadSaveScheduleFunctionality from './LoadSaveFunctionality';
@@ -6,9 +8,34 @@ import Login from './Login';
 import { Logo } from './Logo';
 import AppDrawer from './SettingsMenu';
 
+import trpc from '$lib/api/trpc';
 import { BLUE } from '$src/globals';
+import { useSessionStore } from '$stores/SessionStore';
 
 export function Header() {
+    const { session, updateSession: setSession } = useSessionStore();
+    const [searchParams] = useSearchParams();
+    const navigate = useNavigate();
+    const handleCallback = async () => {
+        try {
+            const code = searchParams.get('code');
+            if (code) {
+                const newSession = await trpc.userData.handleGoogleCallback.mutate({
+                    code: code,
+                    token: session ?? '',
+                });
+
+                setSession(newSession);
+                navigate('/');
+            }
+        } catch (error) {
+            console.error('Error during authentication', error);
+        }
+    };
+
+    useEffect(() => {
+        handleCallback();
+    }, [searchParams]);
     return (
         <AppBar
             position="static"
